+++ R/utils-gh.R
@@ -39,11 +39,14 @@ get_hosturl <- function(url) {
 # (almost) the inverse of get_hosturl()
 # https://github.com     --> https://api.github.com
 # https://github.uni.edu --> https://github.uni.edu/api/v3
+# fmt: skip
 get_apiurl <- function(url) {
   host_url <- get_hosturl(url)
   prot_host <- strsplit(host_url, "://", fixed = TRUE)[[1]]
   if (is_github_dot_com(host_url)) {
     paste0(prot_host[[1]], "://api.github.com")
+  } else if (is_github_enterprise(host_url)) {
+    paste0(prot_host[[1]], "://api.", prot_host[[2]])
   } else {
     paste0(host_url, "/api/v3")
   }
@@ -57,4 +60,13 @@ is_github_dot_com <- function(url) {
 
 default_api_url <- function() {
   Sys.getenv("GITHUB_API_URL", unset = "https://api.github.com")
+}
+
+# handles GitHub Enterprise Cloud, but not GitHub Enterprise Server (which
+# would, I think, require the ability to fully configure this)
+# https://github.com/r-lib/usethis/issues/1897
+is_github_enterprise <- function(url) {
+  url <- get_baseurl(url)
+  url <- normalize_host(url)
+  grepl("^https?://.+ghe.com", url)
 }