+++ NEWS.md
@@ -9,6 +9,8 @@
 * `use_package(min_version = FALSE)` is treated the same as when `min_version`
   is not specified (#2117, @salim-b).
 
+* `use_r_universe_badge()` is a new function that creates a README badge indicating your package is available on [R-universe](https://r-universe.dev) and reporting the latest version (@olivroy, #1883).
+
 * usethis's criteria for recognizing a project have expanded to include (#2133):
   - a `.vscode/settings.json` file, which Positron or VS Code might create
   - a `_quarto.yml` file, typical of a Quarto project
@@ -37,6 +39,11 @@
 * `use_tidy_upkeep_issue()` records the year it is being run in the
   `Config/usethis/upkeep` field in DESCRIPTION. If this value exists, it is
   used to filter the checklist when making the issue.
+
+* Some GitHub functionality should now work for GitHub Enterprise Cloud (as
+  opposed to GitHub Enterprise Server). Specifically an HTTPS URL such as
+  `"https://my-cool-org.ghe.com"` should now be recognized as "looks like
+  GitHub" (#2098, @jameslairdsmith).
 
 # usethis 3.0.0